requirements.txt
requests
dbt-core
python-dotenv
ruff